@@ -1,128 +1,168 @@
-import { EventEmitter } from 'events';
-import * as fs from 'fs';
-import * as path from 'path';
-import * as os from 'os';
-import * as vscode from 'vscode';
+import { EventEmitter } from 'events'
+import * as fs from 'fs'
+import * as path from 'path'
+import * as os from 'os'
+import * as vscode from 'vscode'
 
-import { promisify } from 'util';
+import { promisify } from 'util'
 
-import { formatParam, formatString, showInfo, showError, getUploadedName } from '../utils';
+import {
+  formatParam,
+  formatString,
+  showInfo,
+  showError,
+  getUploadedName,
+  asyncWrapper
+} from '../utils'
+import { IImgInfo, IPlugin, IPicGo, IConfig } from 'picgo/dist/src/types'
 
-const _ = require('lodash');
-const _db = require('lodash-id');
-import nls = require('../../package.nls.json');
-import { IImgInfo, IPlugin, IPicGo, IConfig } from 'picgo/dist/src/types';
-import PicGo = require('picgo');
-_.mixin(_db);
+import _ from '../utils/lodash-mixins'
 
-const writeFileP = promisify(fs.writeFile);
-const readFileP = promisify(fs.readFile);
+import nls = require('../../package.nls.json')
+import PicGo = require('picgo')
+
+const writeFileP = promisify(fs.writeFile)
+const readFileP = promisify(fs.readFile)
 
 export interface INotice {
-  body: string;
-  text: string;
-  title: string;
+  body: string
+  text: string
+  title: string
 }
 
 export interface IUploadName {
-  date: string;
-  dateTime: string;
-  fileName: string;
-  extName: string;
-  mdFileName: string;
-  [key: string]: string;
+  date: string
+  dateTime: string
+  fileName: string
+  extName: string
+  mdFileName: string
+  [key: string]: string
 }
 
 export interface IOutputUrl {
-  uploadedName: string;
-  url: string;
-  [key: string]: string;
+  uploadedName: string
+  url: string
+  [key: string]: string
 }
 
 export enum EVSPicgoHooks {
-  updated = 'updated',
+  updated = 'updated'
 }
 
 export default class VSPicgo extends EventEmitter {
-  private static picgo: IPicGo = new PicGo();
+  private static readonly picgo: IPicGo = new PicGo()
 
   constructor() {
-    super();
-    this.configPicgo();
+    super()
+    this.configPicgo()
     // Before upload, we change names of the images.
-    this.registerRenamePlugin();
+    this.registerRenamePlugin()
     // After upload, we use the custom output format.
-    this.addGenerateOutputListener();
+    this.addGenerateOutputListener()
   }
 
   configPicgo() {
-    const picgoConfigPath = vscode.workspace.getConfiguration('picgo').get<string>('configPath');
-    let config: Partial<IConfig>;
+    const picgoConfigPath = vscode.workspace
+      .getConfiguration('picgo')
+      .get<string>('configPath')
+    let config: Partial<IConfig>
     if (picgoConfigPath) {
       config = JSON.parse(
         fs.readFileSync(picgoConfigPath, {
-          encoding: 'utf-8',
-        }),
-      );
+          encoding: 'utf-8'
+        })
+      )
     } else {
-      const picBed = (vscode.workspace.getConfiguration('picgo.picBed') as any) as IConfig['picBed'];
-      config = { picBed };
+      const picBed = (vscode.workspace.getConfiguration(
+        'picgo.picBed'
+      ) as any) as IConfig['picBed']
+      config = { picBed }
     }
 
     // `PICGO_ENV` is used only by Electron version, should be disabled here,
     // see https://github.com/PicGo/vs-picgo/issues/75 for more detail
-    (config as any)['PICGO_ENV'] = 'CLI';
-    VSPicgo.picgo.setConfig(config);
+    ;(config as any).PICGO_ENV = 'CLI'
+    VSPicgo.picgo.setConfig(config)
   }
 
   addGenerateOutputListener() {
-    VSPicgo.picgo.on('finished', async (ctx: PicGo) => {
-      let urlText = '';
-      const outputFormatTemplate =
-        vscode.workspace.getConfiguration('picgo').get<string>('customOutputFormat') || '![${uploadedName}](${url})';
-      try {
-        urlText = ctx.output.reduce((acc: string, imgInfo: IImgInfo): string => {
-          return `${acc}${formatString(outputFormatTemplate, {
-            uploadedName: getUploadedName(imgInfo),
-            url: imgInfo.imgUrl,
-          })}\n`;
-        }, '');
-        urlText = urlText.trim();
-        await this.updateData(ctx.output);
-      } catch (err) {
-        if (err instanceof SyntaxError) {
-          showError(
-            `the data file ${this.dataPath} has syntax error, ` +
-              `please fix the error by yourself or delete the data file and vs-picgo will recreate for you.`,
-          );
-        } else {
-          showError(`failed to read from data file ${this.dataPath}: ${err || ''}`);
+    VSPicgo.picgo.on(
+      'finished',
+      asyncWrapper(async (ctx: PicGo) => {
+        let urlText = ''
+        const outputFormatTemplate =
+          vscode.workspace
+            .getConfiguration('picgo')
+            // eslint-disable-next-line no-template-curly-in-string
+            .get<string>('customOutputFormat') ?? '![${uploadedName}](${url})'
+        try {
+          urlText = ctx.output.reduce(
+            (acc: string, imgInfo: IImgInfo): string => {
+              return `${acc}${formatString(outputFormatTemplate, {
+                uploadedName: getUploadedName(imgInfo),
+                url: imgInfo.imgUrl
+              })}\n`
+            },
+            ''
+          )
+          urlText = urlText.trim()
+          await this.updateData(ctx.output)
+        } catch (err) {
+          if (err instanceof SyntaxError) {
+            await showError(
+              `the data file ${this.dataPath} has syntax error, ` +
+                `please fix the error by yourself or delete the data file and vs-picgo will recreate for you.`
+            )
+          } else {
+            await showError(
+              `failed to read from data file ${this.dataPath}: ${String(
+                err ?? ''
+              )}`
+            )
+          }
+          return
         }
-        return;
-      }
-      this.editor.edit(textEditor => {
-        textEditor.replace(this.editor.selection, urlText);
-        showInfo(`image uploaded successfully.`);
-        this.emit(EVSPicgoHooks.updated, urlText);
-      });
-    });
+        const editor = this.editor
+        await editor?.edit(
+          asyncWrapper(async (textEditor) => {
+            textEditor.replace(editor.selection, urlText)
+            await showInfo(`image uploaded successfully.`)
+            this.emit(EVSPicgoHooks.updated, urlText)
+          })
+        )
+      })
+    )
   }
 
   registerRenamePlugin() {
-    let beforeUploadPlugin: IPlugin = {
+    const beforeUploadPlugin: IPlugin = {
       handle: (ctx: IPicGo) => {
         const uploadNameTemplate =
-          vscode.workspace.getConfiguration('picgo').get<string>('customUploadName') || '${fileName}';
+          vscode.workspace
+            .getConfiguration('picgo')
+            // eslint-disable-next-line no-template-curly-in-string
+            .get<string>('customUploadName') ?? '${fileName}'
         if (ctx.output.length === 1) {
-          ctx.output[0].fileName = this.changeFilename(ctx.output[0].fileName || '', uploadNameTemplate, undefined);
+          ctx.output[0].fileName = this.changeFilename(
+            ctx.output[0].fileName ?? '',
+            uploadNameTemplate,
+            undefined
+          )
         } else {
           ctx.output.forEach((imgInfo: IImgInfo, index: number) => {
-            imgInfo.fileName = this.changeFilename(imgInfo.fileName || '', uploadNameTemplate, index);
-          });
+            imgInfo.fileName = this.changeFilename(
+              imgInfo.fileName ?? '',
+              uploadNameTemplate,
+              index
+            )
+          })
         }
-      },
-    };
-    VSPicgo.picgo.helper.beforeUploadPlugins.register('vsPicgoRenamePlugin', beforeUploadPlugin);
+      }
+    }
+    VSPicgo.picgo.helper.beforeUploadPlugins.register(
+      'vsPicgoRenamePlugin',
+      beforeUploadPlugin
+    )
   }
 
   /**
@@ -130,89 +170,115 @@ export default class VSPicgo extends EventEmitter {
    * @param original The filename of the original image file.
    * @param template The template string.
    */
-  changeFilename(original: string, template: string, index: number | undefined) {
+  changeFilename(
+    original: string,
+    template: string,
+    index: number | undefined
+  ) {
     if (this.userDefineName) {
-      original = this.userDefineName + (index || '') + path.extname(original);
+      original = `${this.userDefineName}${index ?? ''}${path.extname(original)}`
     }
-    const mdFilePath = this.editor.document.fileName;
-    const mdFileName = path.basename(mdFilePath, path.extname(mdFilePath));
-    let uploadNameData = formatParam(original, mdFileName);
-    return formatString(template, uploadNameData);
+    const mdFilePath = this.editor?.document.fileName
+    if (!mdFilePath) return
+    const mdFileName = path.basename(mdFilePath, path.extname(mdFilePath))
+    const uploadNameData = formatParam(original, mdFileName)
+    return formatString(template, uploadNameData)
   }
 
-  get editor(): vscode.TextEditor {
-    const editor = vscode.window.activeTextEditor;
-    if (!editor) {
-      showError('no active markdown editor!');
+  get editor() {
+    const editor = vscode.window.activeTextEditor
+    if (editor == null) {
+      asyncWrapper(async () => {
+        await showError('no active markdown editor!')
+      })()
+      return null
     }
-    return editor as vscode.TextEditor;
+    return editor
   }
+
   get dataPath(): string {
-    const picgoConfig = vscode.workspace.getConfiguration('picgo');
-    return picgoConfig.dataPath || path.resolve(os.homedir(), 'vs-picgo-data.json');
+    const picgoConfig = vscode.workspace.getConfiguration('picgo')
+    return (
+      picgoConfig.dataPath || path.resolve(os.homedir(), 'vs-picgo-data.json')
+    )
   }
-  get userDefineName(): string {
-    let selectedString = this.editor.document.getText(this.editor.selection);
-    const nameReg = /[:\/\?\$]+/g; // limitations of name
-    selectedString = selectedString.replace(nameReg, () => '');
-    return selectedString;
+
+  get userDefineName() {
+    let selectedString = this.editor?.document.getText(this.editor.selection)
+    const nameReg = /[:/?$]+/g // limitations of name
+    selectedString = selectedString?.replace(nameReg, () => '')
+    return selectedString
   }
 
   async initDataFile(dataPath: string) {
     if (!fs.existsSync(dataPath)) {
-      await writeFileP(dataPath, JSON.stringify({ uploaded: [] }, null, 2), 'utf8');
+      await writeFileP(
+        dataPath,
+        JSON.stringify({ uploaded: [] }, null, 2),
+        'utf8'
+      )
     }
   }
 
   async upload(input?: string[]) {
     // This is necessary, because user may have changed settings
-    this.configPicgo();
+    this.configPicgo()
 
     // uploading progress
-    vscode.window.withProgress(
+    await vscode.window.withProgress(
       {
         location: vscode.ProgressLocation.Notification,
         title: `${nls['ext.displayName']}: image uploading...`,
-        cancellable: false,
+        cancellable: false
       },
-      progress => {
-        return new Promise<void>((resolve, reject) => {
+      async (progress) => {
+        return await new Promise<void>((resolve, reject) => {
           VSPicgo.picgo.on('uploadProgress', (p: number) => {
-            progress.report({ increment: p });
+            progress.report({ increment: p })
             if (p === 100) {
-              resolve();
+              resolve()
             }
-          });
-          VSPicgo.picgo.on('failed', (error: Error) => {
-            showError(error.message || 'Unknown error');
-            reject();
-          });
-          VSPicgo.picgo.on('notification', (notice: INotice) => {
-            showError(`${notice.title}! ${notice.body || ''}${notice.text || ''}`);
-            reject();
-          });
-        });
-      },
-    );
+          })
+          VSPicgo.picgo.on(
+            'failed',
+            asyncWrapper(async (error: Error) => {
+              const errorReason = error.message || 'Unknown error'
+              await showError(errorReason)
+              reject(errorReason)
+            })
+          )
+          VSPicgo.picgo.on(
+            'notification',
+            asyncWrapper(async (notice: INotice) => {
+              const errorReason = `${notice.title}! ${notice.body || ''}${
+                notice.text || ''
+              }`
+              await showError(errorReason)
+              reject(errorReason)
+            })
+          )
+        })
+      }
+    )
 
     // Error has been handled in on 'failed'
-    return VSPicgo.picgo.upload(input).catch(() => {});
+    return await VSPicgo.picgo.upload(input).catch(() => {})
   }
 
-  async updateData(picInfos: Array<IImgInfo>) {
-    const dataPath = this.dataPath;
+  async updateData(picInfos: IImgInfo[]) {
+    const dataPath = this.dataPath
     if (!fs.existsSync(dataPath)) {
-      await this.initDataFile(dataPath);
-      showInfo('data file created at ${dataPath}.');
+      await this.initDataFile(dataPath)
+      await showInfo(`data file created at ${dataPath}.`)
     }
-    const dataRaw = await readFileP(dataPath, 'utf8');
-    const data = JSON.parse(dataRaw);
+    const dataRaw = await readFileP(dataPath, 'utf8')
+    const data = JSON.parse(dataRaw)
     if (!data.uploaded) {
-      data.uploaded = [];
+      data.uploaded = []
     }
-    picInfos.forEach(picInfo => {
-      _.insert(data['uploaded'], picInfo);
-    });
-    await writeFileP(dataPath, JSON.stringify(data, null, 2), 'utf8');
+    picInfos.forEach((picInfo) => {
+      _.insert(data.uploaded, picInfo)
+    })
+    await writeFileP(dataPath, JSON.stringify(data, null, 2), 'utf8')
   }
 }
