@@ -1,7 +1,11 @@
-import * as assert from 'assert';
-import { workspace, window, Selection } from 'vscode';
+/* eslint-disable no-useless-escape */
+/* eslint-disable no-template-curly-in-string */
+/* eslint-disable @typescript-eslint/no-misused-promises */
+/* eslint-disable no-undef */
+import * as assert from 'assert'
+import { workspace, window, Selection } from 'vscode'
 
-import VSPicgo from '../../src/vs-picgo';
+import VSPicgo from '../../src/vs-picgo'
 
 import {
   DEFAULT_CONFIGS,
@@ -9,82 +13,93 @@ import {
   IVSPicgoConfigurationKeys,
   TEST_MD_FILE_PATH,
   TEST_PICTURE_PATH,
-  VSPicgoUploadStarter,
-} from '../utils';
+  VSPicgoUploadStarter
+} from '../utils'
 
-const vspicgo = new VSPicgo();
-let previousConfigs: IVSPicgoConfiguration = Object.assign({}, DEFAULT_CONFIGS);
+const vspicgo = new VSPicgo()
+const previousConfigs: IVSPicgoConfiguration = Object.assign(
+  {},
+  DEFAULT_CONFIGS
+)
 
-const configKeys = Object.keys(previousConfigs);
+const configKeys = Object.keys(previousConfigs)
 
-const REG4CUSTOM_OUTPUT_FORMAT = /^\!\[.+\]\(.+\)$/;
-const REG4CUSTOM_FILE_FORMAT = /^\!\[\d{4}\-\d{2}\-\d{2}]\(.+\)$/;
+const REG4CUSTOM_OUTPUT_FORMAT = /^\!\[.+\]\(.+\)$/
+const REG4CUSTOM_FILE_FORMAT = /^\!\[\d{4}\-\d{2}\-\d{2}]\(.+\)$/
 
 before(async () => {
-  const document = await workspace.openTextDocument(TEST_MD_FILE_PATH);
-  await window.showTextDocument(document);
-});
+  const document = await workspace.openTextDocument(TEST_MD_FILE_PATH)
+  await window.showTextDocument(document)
+})
 
-describe('VSPicgo', async function() {
+describe('VSPicgo', async function () {
   beforeEach(() => {
     // save old configuration info
     for (const section of configKeys) {
-      previousConfigs[section as IVSPicgoConfigurationKeys] = workspace.getConfiguration('', null).get(section);
+      previousConfigs[
+        section as IVSPicgoConfigurationKeys
+      ] = workspace.getConfiguration('', null).get(section)
     }
-  });
+  })
 
-  afterEach(async function() {
+  afterEach(async function () {
     for (const section of configKeys) {
       await workspace
         .getConfiguration('', null)
-        .update(section, previousConfigs[section as IVSPicgoConfigurationKeys], true);
+        .update(
+          section,
+          previousConfigs[section as IVSPicgoConfigurationKeys],
+          true
+        )
     }
-  });
+  })
 
-  it('customOutputFormat should work', async function() {
-    this.timeout(30000);
-    const res = await VSPicgoUploadStarter({
-      args4uploader: [TEST_PICTURE_PATH],
-      configuration: { 'picgo.customOutputFormat': '![${uploadedName}](${url})' },
-      editor: {
-        content: '',
-        selection: new Selection(0, 0, 0, 0),
-      },
-      vspicgo,
-    });
-    console.log('output: ' + res);
-    assert.equal(REG4CUSTOM_OUTPUT_FORMAT.test(res), true);
-  });
-
-  it('customUploadName should work', async function() {
-    this.timeout(30000);
+  it('customOutputFormat should work', async function () {
+    this.timeout(30000)
     const res = await VSPicgoUploadStarter({
       args4uploader: [TEST_PICTURE_PATH],
       configuration: {
-        'picgo.customUploadName': '${date}${extName}',
+        'picgo.customOutputFormat': '![${uploadedName}](${url})'
       },
       editor: {
         content: '',
-        selection: new Selection(0, 0, 0, 0),
+        selection: new Selection(0, 0, 0, 0)
       },
-      vspicgo,
-    });
-    console.log('uploadName: ' + res);
-    assert.equal(REG4CUSTOM_FILE_FORMAT.test(res), true);
-  });
+      vspicgo
+    })
+    console.log('output: ' + res)
+    assert.equal(REG4CUSTOM_OUTPUT_FORMAT.test(res), true)
+  })
 
-  it('selection as fileName should work', async function() {
-    this.timeout(30000);
+  it('customUploadName should work', async function () {
+    this.timeout(30000)
+    const res = await VSPicgoUploadStarter({
+      args4uploader: [TEST_PICTURE_PATH],
+      configuration: {
+        'picgo.customUploadName': '${date}${extName}'
+      },
+      editor: {
+        content: '',
+        selection: new Selection(0, 0, 0, 0)
+      },
+      vspicgo
+    })
+    console.log('uploadName: ' + res)
+    assert.equal(REG4CUSTOM_FILE_FORMAT.test(res), true)
+  })
+
+  it('selection as fileName should work', async function () {
+    this.timeout(30000)
     const res = await VSPicgoUploadStarter({
       args4uploader: [TEST_PICTURE_PATH],
       configuration: {},
       editor: {
         content: 'TEST',
-        selection: new Selection(0, 0, 0, 4),
+        selection: new Selection(0, 0, 0, 4)
       },
-      vspicgo,
-    });
-    console.log('selection: ' + res);
-    assert.equal(res.indexOf('TEST'), 2);
-  });
-});
+      vspicgo
+    })
+    console.log('selection: ' + res)
+    assert.equal(res.indexOf('TEST'), 2)
+  })
+})
