@@ -1,49 +1,56 @@
-import * as fs from 'fs';
-import * as glob from 'glob';
-import * as path from 'path';
-import * as istanbul from 'istanbul';
-import { ICoverageCollectorOptions } from './constants-and-interfaces';
+/* eslint-disable @typescript-eslint/restrict-plus-operands */
+/* eslint-disable @typescript-eslint/no-var-requires */
+import * as fs from 'fs'
+import * as glob from 'glob'
+import * as path from 'path'
+import * as istanbul from 'istanbul'
+import { ICoverageCollectorOptions } from './constants-and-interfaces'
 
-const remapIstanbul = require('remap-istanbul');
+const remapIstanbul = require('remap-istanbul')
 
-declare var global: any;
+declare let global: any
 
 function ensureDirExists(dir: string): void {
   if (!fs.existsSync(dir)) {
-    fs.mkdirSync(dir);
+    fs.mkdirSync(dir)
   }
 }
 
 export class CoverageCollector {
-  private coverageVar: string = '$$cov_' + new Date().getTime() + '$$';
-  private transformer: any;
-  private matchFn: any;
-  private instrumenter: any;
+  private coverageVar: string = '$$cov_' + new Date().getTime() + '$$'
+  private transformer: any
+  private matchFn: any
+  private instrumenter: any
 
-  constructor(basePath: string, private options: ICoverageCollectorOptions) {
+  constructor(
+    basePath: string,
+    private readonly options: ICoverageCollectorOptions
+  ) {
     if (!this.options.sourceRoot) {
-      return;
+      return
     }
-    this.options.sourceRoot = path.join(basePath, this.options.sourceRoot);
-    this.options.output = path.join(basePath, this.options.output);
+    this.options.sourceRoot = path.join(basePath, this.options.sourceRoot)
+    this.options.output = path.join(basePath, this.options.output)
   }
 
   public setup(): void {
     // Set up Code Coverage, hooking require so that instrumented code is returned
-    this.instrumenter = new istanbul.Instrumenter({ coverageVariable: this.coverageVar });
+    this.instrumenter = new istanbul.Instrumenter({
+      coverageVariable: this.coverageVar
+    })
 
     // Glob source files
     const srcFiles = glob.sync('**/**.js', {
       cwd: this.options.sourceRoot,
-      ignore: this.options.ignorePatterns,
-    });
+      ignore: this.options.ignorePatterns
+    })
 
     // Create a match function - taken from the run-with-cover.js in istanbul.
-    const decache = require('decache');
-    const fileMap: any = {};
-    srcFiles.forEach(file => {
-      const fullPath = path.join(this.options.sourceRoot, file);
-      fileMap[fullPath] = true;
+    const decache = require('decache')
+    const fileMap: any = {}
+    srcFiles.forEach((file) => {
+      const fullPath = path.join(this.options.sourceRoot, file)
+      fileMap[fullPath] = true
 
       // On Windows, extension is loaded pre-test hooks and this mean we lose
       // our chance to hook the Require call. In order to instrument the code
@@ -51,21 +58,25 @@ export class CoverageCollector {
       // This doesn't impact tests, but is a concern if we had some integration
       // tests that relied on VSCode accessing our module since there could be
       // some shared global state that we lose.
-      decache(fullPath);
-    });
+      decache(fullPath)
+    })
 
-    this.matchFn = (file: string): boolean => fileMap[file];
-    this.matchFn.files = Object.keys(fileMap);
+    this.matchFn = (file: string): boolean => fileMap[file]
+    this.matchFn.files = Object.keys(fileMap)
 
     // Hook up to the Require function so that when this is called, if any of our source files
     // are required, the instrumented version is pulled in instead. These instrumented versions
     // write to a global coverage variable with hit counts whenever they are accessed
-    this.transformer = this.instrumenter.instrumentSync.bind(this.instrumenter);
-    const hookOpts = { verbose: false, extensions: ['.js'] };
-    (istanbul.hook as any).hookRequire(this.matchFn, this.transformer, hookOpts);
+    this.transformer = this.instrumenter.instrumentSync.bind(this.instrumenter)
+    const hookOpts = { verbose: false, extensions: ['.js'] }
+    ;(istanbul.hook as any).hookRequire(
+      this.matchFn,
+      this.transformer,
+      hookOpts
+    )
 
     // initialize the global variable to stop mocha from complaining about leaks
-    global[this.coverageVar] = {};
+    global[this.coverageVar] = {}
   }
 
   /**
@@ -77,13 +88,18 @@ export class CoverageCollector {
    * @memberOf CoverageRunner
    */
   public report(): void {
-    (istanbul.hook as any).unhookRequire();
-    let cov: any;
-    if (typeof global[this.coverageVar] === 'undefined' || Object.keys(global[this.coverageVar]).length === 0) {
-      console.error('No coverage information was collected, exit without writing coverage information');
-      return;
+    ;(istanbul.hook as any).unhookRequire()
+    let cov: any
+    if (
+      typeof global[this.coverageVar] === 'undefined' ||
+      Object.keys(global[this.coverageVar]).length === 0
+    ) {
+      console.error(
+        'No coverage information was collected, exit without writing coverage information'
+      )
+      return
     } else {
-      cov = global[this.coverageVar];
+      cov = global[this.coverageVar]
     }
 
     // TODO consider putting this under a conditional flag
@@ -91,45 +107,50 @@ export class CoverageCollector {
     // illustrate the missing coverage.
     this.matchFn.files.forEach((file: any) => {
       if (cov[file]) {
-        return;
+        return
       }
-      this.transformer(fs.readFileSync(file, 'utf-8'), file);
+      this.transformer(fs.readFileSync(file, 'utf-8'), file)
 
       // When instrumenting the code, istanbul will give each FunctionDeclaration a value of 1 in coverState.s,
       // presumably to compensate for function hoisting. We need to reset this, as the function was not hoisted,
       // as it was never loaded.
-      Object.keys(this.instrumenter.coverState.s).forEach(key => {
-        this.instrumenter.coverState.s[key] = 0;
-      });
+      Object.keys(this.instrumenter.coverState.s).forEach((key) => {
+        this.instrumenter.coverState.s[key] = 0
+      })
 
-      cov[file] = this.instrumenter.coverState;
-    });
+      cov[file] = this.instrumenter.coverState
+    })
 
     // TODO Allow config of reporting directory with
-    const includePid = this.options.includePid;
-    const pidExt = includePid ? '-' + process.pid : '';
-    const coverageFile = path.resolve(this.options.output, 'coverage' + pidExt + '.json');
+    const includePid = this.options.includePid
+    // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
+    const pidExt = includePid ? '-' + process.pid : ''
+    const coverageFile = path.resolve(
+      this.options.output,
+      'coverage' + pidExt + '.json'
+    )
 
     // yes, do this again since some test runners could clean the dir initially created
-    ensureDirExists(this.options.output);
+    ensureDirExists(this.options.output)
 
-    fs.writeFileSync(coverageFile, JSON.stringify(cov), 'utf8');
+    fs.writeFileSync(coverageFile, JSON.stringify(cov), 'utf8')
 
     const remappedCollector = remapIstanbul.remap(cov, {
       warn: (warning: any) => {
         // We expect some warnings as any JS file without a typescript mapping will cause this.
         // By default, we'll skip printing these to the console as it clutters it up
         if (this.options.verbose) {
-          console.warn(warning);
+          console.warn(warning)
         }
-      },
-    });
+      }
+    })
 
-    const reporter = new istanbul.Reporter(undefined, this.options.output);
-    const reportTypes = this.options.reports instanceof Array ? this.options.reports : ['lcov'];
-    reporter.addAll(reportTypes);
+    const reporter = new istanbul.Reporter(undefined, this.options.output)
+    const reportTypes =
+      this.options.reports instanceof Array ? this.options.reports : ['lcov']
+    reporter.addAll(reportTypes)
     reporter.write(remappedCollector, true, () => {
-      console.log(`reports written to ${this.options.output}`);
-    });
+      console.log(`reports written to ${this.options.output}`)
+    })
   }
 }
