@@ -1,3 +1,3 @@
-export * from './constants-and-interfaces';
-export * from './coverage-collector';
-export * from './upload-starter';
+export * from './constants-and-interfaces'
+export * from './coverage-collector'
+export * from './upload-starter'
